+++ src/compute_shader.rs
@@ -11,7 +11,7 @@ pub struct ComputeShader {
     group_sizes: [u32; 3],
 }
 
-impl VkInit {
+impl<'a> VkInit<'a> {
     /// Only SSBOs are supported as bindings.
     ///
     /// Group sizes are read in as specialization constants: layout(local_size_x_id = 0, local_size_y_id = 1, local_size_z_id = 2) in;
@@ -24,13 +24,9 @@ impl VkInit {
         additional_spec_consts: &[u32],
         base_debug_name: String,
     ) -> Result<ComputeShader, Error> {
-        let module_info = ShaderModuleCreateInfo::builder().code(&code);
+        let module_info = ShaderModuleCreateInfo::default().code(&code);
         let module = unsafe { self.device.create_shader_module(&module_info, None) }?;
-        self.set_debug_object_name(
-            module.as_raw(),
-            ObjectType::SHADER_MODULE,
-            format!("{base_debug_name}_Compute_Shader_Module"),
-        )?;
+        self.set_debug_object_name(module, format!("{base_debug_name}_Compute_Shader_Module"))?;
 
         let mut spec_consts_data: Vec<u8> = vec![];
         spec_consts_data.extend(group_sizes[0].to_ne_bytes());
@@ -50,43 +46,36 @@ impl VkInit {
             })
             .collect();
 
-        let spec_consts_info = SpecializationInfo::builder()
+        let spec_consts_info = SpecializationInfo::default()
             .map_entries(&spec_consts_map_entries)
             .data(&spec_consts_data);
 
         let shader_entry_name = CString::new("main")?;
-        let shader_stage_info = PipelineShaderStageCreateInfo::builder()
+        let shader_stage_info = PipelineShaderStageCreateInfo::default()
             .stage(ShaderStageFlags::COMPUTE)
             .module(module)
             .specialization_info(&spec_consts_info)
             .name(&shader_entry_name);
 
-        let push_constants_ranges = [PushConstantRange::builder()
+        let push_constants_ranges = [PushConstantRange::default()
             .offset(0)
             .size(size_of::<Push>() as u32)
-            .stage_flags(ShaderStageFlags::COMPUTE)
-            .build()];
+            .stage_flags(ShaderStageFlags::COMPUTE)];
 
-        let pool_size = DescriptorPoolSize::builder()
+        let pool_size = DescriptorPoolSize::default()
             .ty(DescriptorType::STORAGE_BUFFER)
-            .descriptor_count(ssbos.len() as u32)
-            .build();
+            .descriptor_count(ssbos.len() as u32);
 
         let pool_sizes = [pool_size];
-        let desc_pool_create_info = DescriptorPoolCreateInfo::builder()
+        let desc_pool_create_info = DescriptorPoolCreateInfo::default()
             .max_sets(1)
-            .pool_sizes(&pool_sizes)
-            .build();
+            .pool_sizes(&pool_sizes);
 
         let desc_pool = unsafe {
             self.device
                 .create_descriptor_pool(&desc_pool_create_info, None)
         }?;
-        self.set_debug_object_name(
-            desc_pool.as_raw(),
-            ObjectType::DESCRIPTOR_POOL,
-            format!("{base_debug_name}_Descriptor_Pool"),
-        )?;
+        self.set_debug_object_name(desc_pool, format!("{base_debug_name}_Descriptor_Pool"))?;
 
         let mut layout_bindings: Vec<DescriptorSetLayoutBinding> = Vec::new();
         let mut descriptor_buffers: Vec<DescriptorBufferInfo> = Vec::new();
@@ -109,33 +98,26 @@ impl VkInit {
             descriptor_buffers.push(descriptor_buffer);
         }
 
-        let desc_set_layout_info = DescriptorSetLayoutCreateInfo::builder()
-            .bindings(&layout_bindings)
-            .build();
+        let desc_set_layout_info =
+            DescriptorSetLayoutCreateInfo::default().bindings(&layout_bindings);
 
         let desc_set_layout = unsafe {
             self.device
                 .create_descriptor_set_layout(&desc_set_layout_info, None)?
         };
         self.set_debug_object_name(
-            desc_set_layout.as_raw(),
-            ObjectType::DESCRIPTOR_SET_LAYOUT,
+            desc_set_layout,
             format!("{base_debug_name}_SSBO_Desc_Layout"),
         )?;
 
         let desc_set_layouts = [desc_set_layout];
-        let alloc_info = ash::vk::DescriptorSetAllocateInfo::builder()
+        let alloc_info = ash::vk::DescriptorSetAllocateInfo::default()
             .descriptor_pool(desc_pool)
-            .set_layouts(&desc_set_layouts)
-            .build();
+            .set_layouts(&desc_set_layouts);
 
         let desc_sets = unsafe { self.device.allocate_descriptor_sets(&alloc_info)? };
         for (i, set) in desc_sets.iter().enumerate() {
-            self.set_debug_object_name(
-                set.as_raw(),
-                ObjectType::DESCRIPTOR_SET,
-                format!("{base_debug_name}_SSBO_Set_{i}"),
-            )?;
+            self.set_debug_object_name(*set, format!("{base_debug_name}_SSBO_Set_{i}"))?;
         }
 
         let mut write_sets: Vec<ash::vk::WriteDescriptorSet> = Vec::new();
@@ -155,40 +137,34 @@ impl VkInit {
             self.device.update_descriptor_sets(&write_sets, &[]);
         }
 
-        let pipeline_layout_info = PipelineLayoutCreateInfo::builder()
+        let pipeline_layout_info = PipelineLayoutCreateInfo::default()
             .set_layouts(&desc_set_layouts)
-            .push_constant_ranges(&push_constants_ranges)
-            .build();
+            .push_constant_ranges(&push_constants_ranges);
 
         let pipeline_layout = unsafe {
             self.device
                 .create_pipeline_layout(&pipeline_layout_info, None)?
         };
         self.set_debug_object_name(
-            pipeline_layout.as_raw(),
-            ObjectType::PIPELINE_LAYOUT,
+            pipeline_layout,
             format!("{base_debug_name}_Pipeline_Layout"),
         )?;
 
-        let pipeline_info = ComputePipelineCreateInfo::builder()
-            .stage(*shader_stage_info)
+        let pipeline_info = ComputePipelineCreateInfo::default()
+            .stage(shader_stage_info)
             .layout(pipeline_layout);
 
         let pipeline = unsafe {
             match self.device.create_compute_pipelines(
                 PipelineCache::null(),
-                &[*pipeline_info],
+                &[pipeline_info],
                 None,
             ) {
                 Ok(pipeline) => pipeline[0],
                 Err((_, e)) => return Err(e.into()),
             }
         };
-        self.set_debug_object_name(
-            pipeline.as_raw(),
-            ObjectType::PIPELINE,
-            format!("{base_debug_name}_Pipeline"),
-        )?;
+        self.set_debug_object_name(pipeline, format!("{base_debug_name}_Pipeline"))?;
 
         Ok(ComputeShader {
             pipeline,